requirements.txt
six~=1.16.0
pillow~=10.4.0
numpy~=1.26.4
yaml~=0.2.5
pyyaml~=6.0.1
optuna~=4.0.0
progress~=1.5
matplotlib~=3.9.2
click~=8.1.7
fsspec~=2024.6.1
pytorch
torchvision
torchaudio
pytorch-cuda~=12.4